wiki draft api

+++ backend/api/v1/views/wiki.py
@@ -15,7 +15,8 @@ from webargs.flaskparser import use_args, use_kwargs
 from backend.api.v1 import api_v1
 from backend.extensions import db
 from backend.models import Category, Post
-from backend.api.v1.schemas import category_schema, categorys_schema, post_schema, posts_schema, post_detail_schema
+from backend.api.v1.schemas import category_schema, categorys_schema, post_schema, posts_schema, post_detail_schema, \
+    drafts_schema
 from backend.decorators import auth_required
 from backend.utils.utils import api_abort, validate_category_id
 
@@ -112,7 +113,7 @@ class CategorysAPI(MethodView):
 
 
 class PostAPI(MethodView):
-    # decorators = [auth_required]
+    decorators = [auth_required]
 
     def get(self, post_id):
         """
@@ -132,6 +133,9 @@ class PostAPI(MethodView):
         :return:
         """
         doc = Post.query.get_or_404(post_id)
+        # 作者本人才可编辑
+        if g.user.id != doc.author.id:
+            return api_abort(403, "您无法编辑此文章")
         content = args["body"]
         soup = filter_html(content)
         # 构建摘要数据,获取标签字符串的文本前150个符号
@@ -142,8 +146,8 @@ class PostAPI(MethodView):
         doc.desc = desc
         doc.body = str(soup)
         doc.update_time = datetime.datetime.now()
-        # doc.author = g.user
-        doc.author_id = 20
+        doc.author = g.user
+        # doc.author_id = 20
         try:
             db.session.add(doc)
             db.session.commit()
@@ -156,6 +160,9 @@ class PostAPI(MethodView):
     def delete(self, post_id):
         """删除文章接口"""
         doc = Post.query.get_or_404(post_id)
+        # 文章作者有删除权限
+        if g.user.id != doc.author.id:
+            return api_abort(403, "无法删除此文档")
         db.session.delete(doc)
         db.session.commit()
         return '', 204
@@ -172,12 +179,17 @@ class PostsAPI(MethodView):
         page = request.args.get('page', 1, type=int)
         limit = request.args.get('limit', type=int)
         category_id = request.args.get('category_id', type=int)
+        author_id = request.args.get('author_id', type=int)
+        # 发布状态
+        published = request.args.get('published', True)
 
         per_page = limit or current_app.config['BACK_ITEM_PER_PAGE']
         pagination = Post.query.filter(
             # 查询搜索条件
-            Post.category_id == category_id if category_id else text('')
-        ).order_by(text('-id')).paginate(page, per_page)
+            Post.published == published,
+            Post.category_id == category_id if category_id else text(''),
+            Post.author_id == author_id if author_id else text('')
+        ).order_by(text('-update_time')).paginate(page, per_page)
         items = pagination.items
         current = url_for('.options', page=page, _external=True)
         prev = None
@@ -205,8 +217,64 @@ class PostsAPI(MethodView):
         doc.category_id = args["category_id"] or 1
         doc.desc = desc
         doc.body = str(soup)
+        doc.published = True
         doc.author = g.user
         # doc.author_id = 20
+        try:
+            db.session.add(doc)
+            db.session.commit()
+        except Exception as e:
+            db.session.rollback()
+            current_app.logger.error(e)
+            return api_abort(400, "数据保存失败")
+        response = jsonify(post_schema(doc))
+        response.status_code = 201
+        return response
+
+
+class DraftAPI(MethodView):
+    decorators = [auth_required]
+
+    def delete(self, draft_id):
+        draft = Post.query.filter(Post.id == draft_id, Post.published == False).first()
+        if not draft: return api_abort(404)
+        # 文章作者有删除权限
+        if g.user.id != draft.author.id:
+            return api_abort(403, "无法删除此文档")
+        db.session.delete(draft)
+        db.session.commit()
+        return '', 204
+
+
+@api_v1.route('/delete-drafts', methods=['post'])
+def batch_draft_del():
+    pass
+
+
+class DraftsAPI(MethodView):
+    decorators = [auth_required]
+
+    def get(self):
+        """获取用户草稿列表接口"""
+        drafts = Post.query.filter(Post.author == g.user, Post.published == False).order_by(text('-update_time')).all()
+        return jsonify(drafts_schema(drafts))
+
+    @use_args(post_args, location="json")
+    def post(self, args):
+        draft_id = request.args.get('draft_id', type=int)
+        doc = Post.query.get(draft_id)
+        if doc:
+            if doc.published == True: return api_abort(403, "该文章已发布")
+        else:
+            doc = Post()
+        content = args["body"]
+        soup = filter_html(content)
+
+        doc.title = args["title"]
+        doc.body = str(soup)
+        doc.update_time = datetime.datetime.now()
+        doc.category_id = args["category_id"] or 1
+        doc.author = g.user
         try:
             db.session.add(doc)
             db.session.commit()
@@ -225,3 +293,5 @@ api_v1.add_url_rule('/category/<int:category_id>', view_func=CategoryAPI.as_view
 api_v1.add_url_rule('/categorys', view_func=CategorysAPI.as_view('categorys'), methods=['GET', 'POST'])
 api_v1.add_url_rule('/post/<int:post_id>', view_func=PostAPI.as_view('post'), methods=['GET', 'PUT', 'DELETE'])
 api_v1.add_url_rule('/posts', view_func=PostsAPI.as_view('posts'), methods=['GET', 'POST'])
+api_v1.add_url_rule('/drafts', view_func=DraftsAPI.as_view('drafts'), methods=['GET', 'POST'])
+api_v1.add_url_rule('/draft/<int:draft_id>', view_func=DraftAPI.as_view('draft'), methods=['DELETE', ])
